adding implementation through ch 5.2
implement the GenerateAst.java file which generates the Expr.java file.
The idea is that the derivations of the Expr production in the Lox
grammar will be implemented as derived classes of a base Expr class. The
derivation implementations all require a similar boilerplate code, so
the GenerateAst.java file handles this boilerplate generation for us.